wip: making the widget
